refactor(player): read player state from context in stage content
